+++ Pod/Classes/MWZMapOptions.m
@@ -10,7 +10,7 @@
 }
 
 - (NSString *)description {
-    return [NSString stringWithFormat: @"MWZMapOptions: ApiKey=%@ MaxBounds=%@ Center=%@ Zoom=%@ Floor=%@ LocationEnabled=%@ BeaconsEnabled=%@", _apiKey, _maxBounds, _center, _zoom, _floor, _locationEnabled ? @"YES" : @"NO", _beaconsEnabled ? @"YES" : @"NO"];
+    return [NSString stringWithFormat: @"MWZMapOptions: ApiKey=%@ MaxBounds=%@ MinZoom=%@Center=%@ Zoom=%@ Floor=%@ LocationEnabled=%@ BeaconsEnabled=%@", _apiKey, _maxBounds, _minZoom, _center, _zoom, _floor, _locationEnabled ? @"YES" : @"NO", _beaconsEnabled ? @"YES" : @"NO"];
 }
 
 @end