+++ Pod/Classes/MWZMapView.m
@@ -4,8 +4,8 @@
 #import "MWZMapOptions.h"
 
 #define SERVER_URL @"https://www.mapwize.io"
-#define SDK_VERSION @"1.6.x"
-#define IOS_SDK_VERSION @"1.6.0"
+#define SDK_VERSION @"1.7.x"
+#define IOS_SDK_VERSION @"1.7.1"
 #define IOS_SDK_NAME @"IOS SDK"
 
 @implementation MWZMapView {
@@ -18,6 +18,7 @@
     MWZMeasurement* _userPosition;
     NSMutableArray* _jsQueue;
     MWZLatLon* _center;
+    NSArray* monitoredUuids;
     MWZMapOptions* _options;
     NSMutableDictionary* callbackMemory;
 }
@@ -26,8 +27,6 @@
     _isWebviewLoaded = NO;
     _jsQueue = [[NSMutableArray alloc] init];
     callbackMemory = [[NSMutableDictionary alloc] init];
-    
-    
     
     /*
      * Loads the webview
@@ -76,24 +75,43 @@
     if (options.language != nil) {
         [optionsDic setObject:options.language forKey:@"language"];
     }
+    if (options.minZoom != nil) {
+        [optionsDic setObject:options.minZoom forKey:@"minZoom"];
+    }
     [optionsDic setObject:@0 forKey:@"useBrowserLocation"];
     [optionsDic setObject:@0 forKey:@"zoomControl"];
     
     NSData *optionsJson = [NSJSONSerialization dataWithJSONObject:optionsDic options:(NSJSONWritingOptions) 0 error:nil];
     NSString* optionsString = [[NSString alloc] initWithData:optionsJson encoding:NSUTF8StringEncoding];
-    
     /*
      * Set up the map with the options
      */
-    [self executeJS:[NSString stringWithFormat:@"var map = Mapwize.map('map',%@, window.webkit.messageHandlers.MWZMapEvent.postMessage({type:'maploaded'}));",optionsString]];
-    
-    
+    NSMutableString* js = [[NSMutableString alloc] init];
+    [js appendString:@"map.on('zoomend', function(e){window.webkit.messageHandlers.MWZMapEvent.postMessage({type:e.type, zoom:map.getZoom()});});"];
+    [js appendString:@"map.on('click', function(e){window.webkit.messageHandlers.MWZMapEvent.postMessage({type:e.type, latlng:e.latlng});});"];
+    [js appendString:@"map.on('contextmenu', function(e){window.webkit.messageHandlers.MWZMapEvent.postMessage({type:e.type, latlng:e.latlng});});"];
+    [js appendString:@"map.on('floorsChange', function(e){window.webkit.messageHandlers.MWZMapEvent.postMessage({type:e.type, floors:this._floors});});"];
+    [js appendString:@"map.on('floorChange', function(e){window.webkit.messageHandlers.MWZMapEvent.postMessage({type:e.type, floor:this._floor});});"];
+    [js appendString:@"map.on('placeClick', function(e){window.webkit.messageHandlers.MWZMapEvent.postMessage({type:e.type, place:e.place});});"];
+    [js appendString:@"map.on('venueClick', function(e){window.webkit.messageHandlers.MWZMapEvent.postMessage({type:e.type, venue:e.venue});});"];
+    [js appendString:@"map.on('markerClick', function(e){window.webkit.messageHandlers.MWZMapEvent.postMessage({type:e.type, lat:e.latlng.lat, lon:e.latlng.lng, floor:e.floor});});"];
+    [js appendString:@"map.on('moveend', function(e){window.webkit.messageHandlers.MWZMapEvent.postMessage({type:e.type, center:map.getCenter()});});"];
+    [js appendString:@"map.on('userPositionChange', function(e){window.webkit.messageHandlers.MWZMapEvent.postMessage({type:e.type, userPosition:e.userPosition});});"];
+    [js appendString:@"map.on('followUserModeChange', function(e){window.webkit.messageHandlers.MWZMapEvent.postMessage({type:e.type, followUserMode:e.active});});"];
+    [js appendString:@"map.on('directionsStart', function(e){window.webkit.messageHandlers.MWZMapEvent.postMessage({type:e.type, info:'Directions have been loaded'});});"];
+    [js appendString:@"map.on('directionsStop', function(e){window.webkit.messageHandlers.MWZMapEvent.postMessage({type:e.type, info:'Directions have stopped'});});"];
+    [js appendString:@"map.on('apiResponse', function(e){window.webkit.messageHandlers.MWZMapEvent.postMessage({type:e.type, returnedType: e.returnedType, hash:e.hash, response:e.response, error:e.error});});"];
+    [js appendString:@"Mapwize.Location.on('monitoredUuidsChange', function(e){window.webkit.messageHandlers.MWZMapEvent.postMessage({type:'monitoredUuidsChange', uuids:e.uuids});});"];
+    [js appendString:@"Mapwize.Location.setUseBrowserLocation(false);"]
+    ;
+    [self executeJS:[NSString stringWithFormat:@"var map = Mapwize.map('map',%@, function () {%@window.webkit.messageHandlers.MWZMapEvent.postMessage({type:'maploaded'})});",optionsString, js]];
 
     /*
      * Configures Location manager (authorizations need to be requested outside the SDK)
      */
     _locationManager = [[CLLocationManager alloc] init];
     _locationManager.delegate = self;
+    monitoredUuids = [[NSArray alloc] init];
     
 }
 
@@ -111,14 +129,12 @@
     } else {
         [_jsQueue addObject:js];
     }
-    
 }
 
 /*
  * Handle the full load of the webview. Any command that was queued so far is executed.
  */
 - (void)webView:(WKWebView *)_webView didFinishNavigation:(WKNavigation *)navigation {
-    
     while ([_jsQueue count] > 0) {
         NSString* js = [_jsQueue objectAtIndex:0];
         [_jsQueue removeObjectAtIndex:0];
@@ -130,14 +146,8 @@
                 }
             }
         }];
-        
     }
-    
     _isWebviewLoaded = YES;
-    
-    if ([self.delegate respondsToSelector:@selector(map:webViewDidFinishLoad:)]) {
-        [self.delegate map:self webViewDidFinishLoad:_webView];
-    }
 }
 
 
@@ -160,31 +170,12 @@
 - (void)userContentController:(WKUserContentController *)userContentController didReceiveScriptMessage:(WKScriptMessage *)message {
     
     NSDictionary* body = message.body;
+    
     if ([body[@"type"] isEqualToString:@"maploaded"]) {
-        /*
-         * Register the event handlers
-         */
-        [self executeJS:@"map.on('zoomend', function(e){window.webkit.messageHandlers.MWZMapEvent.postMessage({type:e.type, zoom:map.getZoom()});});"];
-        [self executeJS:@"map.on('click', function(e){window.webkit.messageHandlers.MWZMapEvent.postMessage({type:e.type, latlng:e.latlng});});"];
-        [self executeJS:@"map.on('contextmenu', function(e){window.webkit.messageHandlers.MWZMapEvent.postMessage({type:e.type, latlng:e.latlng});});"];
-        [self executeJS:@"map.on('floorsChange', function(e){window.webkit.messageHandlers.MWZMapEvent.postMessage({type:e.type, floors:this._floors});});"];
-        [self executeJS:@"map.on('floorChange', function(e){window.webkit.messageHandlers.MWZMapEvent.postMessage({type:e.type, floor:this._floor});});"];
-        [self executeJS:@"map.on('placeClick', function(e){window.webkit.messageHandlers.MWZMapEvent.postMessage({type:e.type, place:e.place});});"];
-        [self executeJS:@"map.on('venueClick', function(e){window.webkit.messageHandlers.MWZMapEvent.postMessage({type:e.type, venue:e.venue});});"];
-        [self executeJS:@"map.on('markerClick', function(e){window.webkit.messageHandlers.MWZMapEvent.postMessage({type:e.type, lat:e.latlng.lat, lon:e.latlng.lng, floor:e.floor});});"];
-        [self executeJS:@"map.on('moveend', function(e){window.webkit.messageHandlers.MWZMapEvent.postMessage({type:e.type, center:map.getCenter()});});"];
-        [self executeJS:@"map.on('userPositionChange', function(e){window.webkit.messageHandlers.MWZMapEvent.postMessage({type:e.type, userPosition:e.userPosition});});"];
-        [self executeJS:@"map.on('followUserModeChange', function(e){window.webkit.messageHandlers.MWZMapEvent.postMessage({type:e.type, followUserMode:e.active});});"];
-        [self executeJS:@"map.on('directionsStart', function(e){window.webkit.messageHandlers.MWZMapEvent.postMessage({type:e.type, info:'Directions have been loaded'});});"];
-        [self executeJS:@"map.on('directionsStop', function(e){window.webkit.messageHandlers.MWZMapEvent.postMessage({type:e.type, info:'Directions have stopped'});});"];
-        [self executeJS:@"map.on('apiResponse', function(e){window.webkit.messageHandlers.MWZMapEvent.postMessage({type:e.type, returnedType: e.returnedType, hash:e.hash, response:e.response, error:e.error});});"];
-        if (_options.beaconsEnabled) {
-            [self executeJS:@"map.on('monitoredUuidsChange', function(e){window.webkit.messageHandlers.MWZMapEvent.postMessage({type:e.type, uuids:e.uuids});});"];
-        }
+        
         if (_options.locationEnabled) {
-            [_locationManager startUpdatingLocation];
+            [self startLocationWithBeacons:_options.beaconsEnabled];
         }
-
         if ([self.delegate respondsToSelector:@selector(mapDidLoad:)]) {
             [self.delegate mapDidLoad: self];
         }
@@ -280,8 +271,8 @@
         }
     }
     else if ([body[@"type"] isEqualToString:@"monitoredUuidsChange"]) {
-        NSArray* uuids = body[@"uuids"];
-        [ self updateMonitoring: uuids ];
+        monitoredUuids = body[@"uuids"];
+        [ self updateMonitoring: monitoredUuids ];
     }
     else if ([body[@"type"] isEqualToString:@"apiResponse"]) {
         NSString* hash = body[@"hash"];
@@ -506,15 +497,15 @@
 }
 
 - (void) newUserPositionMeasurement: (MWZMeasurement*) measurement {
-    [self executeJS:[NSString stringWithFormat:@"map.newUserPositionMeasurement(%@)", [measurement toStringJSON] ]];
+    [self executeJS:[NSString stringWithFormat:@"Mapwize.Location.newUserPositionMeasurement(%@)", [measurement toStringJSON] ]];
 }
 
 - (void) setUserHeading: (NSNumber*) heading {
-    [self executeJS:[NSString stringWithFormat:@"map.setUserHeading(%@)", heading!=nil?heading:@"null" ]];
+    [self executeJS:[NSString stringWithFormat:@"Mapwize.Location.setUserHeading(%@)", heading!=nil?heading:@"null" ]];
 }
 
 - (void) removeUserPosition {
-    [self executeJS:[NSString stringWithFormat:@"map.setUserPosition(null)"]];
+    [self executeJS:[NSString stringWithFormat:@"Mapwize.Location.setUserPosition(null)"]];
 }
 
 - (void) setUserPositionWithLatitude: (NSNumber*) latitude longitude:(NSNumber*) longitude floor:(NSNumber*) floor {
@@ -533,10 +524,10 @@
         [positionDic setObject:@0 forKey:@"accuracy"];
         NSData *userPositionJSON = [NSJSONSerialization dataWithJSONObject:positionDic options:(NSJSONWritingOptions) 0 error:nil];
         NSString* userPositionString = [[NSString alloc] initWithData:userPositionJSON encoding:NSUTF8StringEncoding];
-        [self executeJS:[NSString stringWithFormat:@"map.setUserPosition(%@)", userPositionString ]];
+        [self executeJS:[NSString stringWithFormat:@"Mapwize.Location.setUserPosition(%@)", userPositionString ]];
     }
     else {
-        [self executeJS:[NSString stringWithFormat:@"map.setUserPosition(null)" ]];
+        [self executeJS:[NSString stringWithFormat:@"Mapwize.Location.setUserPosition(null)" ]];
     }
     
 }
@@ -559,15 +550,15 @@
         [positionDic setObject:@0 forKey:@"accuracy"];
         NSData *userPositionJSON = [NSJSONSerialization dataWithJSONObject:positionDic options:(NSJSONWritingOptions) 0 error:nil];
         NSString* userPositionString = [[NSString alloc] initWithData:userPositionJSON encoding:NSUTF8StringEncoding];
-        [self executeJS:[NSString stringWithFormat:@"map.setUserPosition(%@)", userPositionString ]];
+        [self executeJS:[NSString stringWithFormat:@"Mapwize.Location.setUserPosition(%@)", userPositionString ]];
     }
     else {
-        [self executeJS:[NSString stringWithFormat:@"map.setUserPosition(null)" ]];
+        [self executeJS:[NSString stringWithFormat:@"Mapwize.Location.setUserPosition(null)" ]];
     }
 }
 
 - (void) unlockUserPosition {
-    [self executeJS:[NSString stringWithFormat:@"map.unlockUserPosition()"]];
+    [self executeJS:[NSString stringWithFormat:@"Mapwize.Location.unlockUserPosition()"]];
     
 }
 
@@ -636,7 +627,7 @@
     void(^_handler)(BOOL);
     _handler = [handler copy];
     [callbackMemory setValue:_handler forKey:hash];
-    [self executeJS:[NSString stringWithFormat:@"map.access('%@', function(isValid){map.fire('apiResponse', {returnedType:'access', hash:'%@', response:isValid});});", accessKey, hash ]];
+    [self executeJS:[NSString stringWithFormat:@"map.access('%@', function(isValid){map.fire('apiResponse', {returnedType:'access', hash:'%@', response:isValid}, true);});", accessKey, hash ]];
 }
 
 /* Margin */
@@ -667,8 +658,7 @@
     _handler = [handler copy];
     [callbackMemory setValue:_handler forKey:hash];
     
-    [self executeJS:[NSString stringWithFormat:@"Mapwize.api.getPlace('%@', function(err, place){map.fire('apiResponse', {returnedType:'place', hash:'%@', response:place, error:err?err.status:''});});", placeId, hash ]];
-    
+    [self executeJS:[NSString stringWithFormat:@"Mapwize.Api.getPlace('%@', function(err, place){map.fire('apiResponse', {returnedType:'place', hash:'%@', response:place, error:err?err.status:''});});", placeId, hash ]];
 }
 
 - (void) getPlaceWithAlias: (NSString*) placeAlias inVenue: (NSString*) venueId completionHandler:(void(^)(MWZPlace*, NSError*)) handler {
@@ -677,7 +667,7 @@
     _handler = [handler copy];
     [callbackMemory setValue:_handler forKey:hash];
     
-    [self executeJS:[NSString stringWithFormat:@"Mapwize.api.getPlace({alias:'%@', venueId:'%@'}, function(err, place){map.fire('apiResponse', {returnedType:'place', hash:'%@', response:place, error:err?err.status:''});});", placeAlias, venueId, hash ]];
+    [self executeJS:[NSString stringWithFormat:@"Mapwize.Api.getPlace({alias:'%@', venueId:'%@'}, function(err, place){map.fire('apiResponse', {returnedType:'place', hash:'%@', response:place, error:err?err.status:''});});", placeAlias, venueId, hash ]];
 }
 
 - (void) getPlaceWithName: (NSString*) placeName inVenue: (NSString*) venueId completionHandler:(void(^)(MWZPlace*, NSError*)) handler {
@@ -686,7 +676,7 @@
     _handler = [handler copy];
     [callbackMemory setValue:_handler forKey:hash];
     
-    [self executeJS:[NSString stringWithFormat:@"Mapwize.api.getPlace({name:'%@', venueId:'%@'}, function(err, place){map.fire('apiResponse', {returnedType:'place', hash:'%@', response:place, error:err?err.status:''});});", placeName, venueId, hash ]];
+    [self executeJS:[NSString stringWithFormat:@"Mapwize.Api.getPlace({name:'%@', venueId:'%@'}, function(err, place){map.fire('apiResponse', {returnedType:'place', hash:'%@', response:place, error:err?err.status:''});});", placeName, venueId, hash ]];
 }
 
 - (void) getVenueWithId: (NSString*) venueId completionHandler:(void(^)(MWZVenue*, NSError*)) handler {
@@ -695,7 +685,7 @@
     _handler = [handler copy];
     [callbackMemory setValue:_handler forKey:hash];
 
-    [self executeJS:[NSString stringWithFormat:@"Mapwize.api.getVenue('%@', function(err, venue){map.fire('apiResponse', {returnedType:'venue', hash:'%@', response:venue, error:err?err.status:''});});", venueId, hash ]];
+    [self executeJS:[NSString stringWithFormat:@"Mapwize.Api.getVenue('%@', function(err, venue){map.fire('apiResponse', {returnedType:'venue', hash:'%@', response:venue, error:err?err.status:''});});", venueId, hash ]];
 }
 
 - (void) getVenueWithName: (NSString*) venueName completionHandler:(void(^)(MWZVenue*, NSError*)) handler {
@@ -704,7 +694,7 @@
     _handler = [handler copy];
     [callbackMemory setValue:_handler forKey:hash];
     
-    [self executeJS:[NSString stringWithFormat:@"Mapwize.api.getVenue({venueName:'%@'}, function(err, venue){map.fire('apiResponse', {returnedType:'venue', hash:'%@', response:venue, error:err?err.status:''});});", venueName, hash ]];
+    [self executeJS:[NSString stringWithFormat:@"Mapwize.Api.getVenue({venueName:'%@'}, function(err, venue){map.fire('apiResponse', {returnedType:'venue', hash:'%@', response:venue, error:err?err.status:''});});", venueName, hash ]];
 }
 
 - (void) getVenueWithAlias: (NSString*) venueAlias completionHandler:(void(^)(MWZVenue*, NSError*)) handler {
@@ -713,7 +703,7 @@
     _handler = [handler copy];
     [callbackMemory setValue:_handler forKey:hash];
     
-    [self executeJS:[NSString stringWithFormat:@"Mapwize.api.getVenue({venueAlias:'%@'}, function(err, venue){map.fire('apiResponse', {returnedType:'venue', hash:'%@', response:venue, error:err?err.status:''});});", venueAlias, hash ]];
+    [self executeJS:[NSString stringWithFormat:@"Mapwize.Api.getVenue({venueAlias:'%@'}, function(err, venue){map.fire('apiResponse', {returnedType:'venue', hash:'%@', response:venue, error:err?err.status:''});});", venueAlias, hash ]];
 }
 
 - (void) getPlaceListWithId: (NSString*) placeListId completionHandler:(void(^)(MWZPlaceList*, NSError*)) handler {
@@ -722,7 +712,7 @@
     _handler = [handler copy];
     [callbackMemory setValue:_handler forKey:hash];
     
-    [self executeJS:[NSString stringWithFormat:@"Mapwize.api.getPlaceList('%@', function(err, placeList){map.fire('apiResponse', {returnedType:'placeList', hash:'%@', response:placeList, error:err?err.status:''});});", placeListId, hash ]];
+    [self executeJS:[NSString stringWithFormat:@"Mapwize.Api.getPlaceList('%@', function(err, placeList){map.fire('apiResponse', {returnedType:'placeList', hash:'%@', response:placeList, error:err?err.status:''});});", placeListId, hash ]];
 }
 
 - (void) getPlaceListWithName: (NSString*) placeListName inVenue:(NSString*) venueId completionHandler:(void(^)(MWZPlaceList*, NSError*)) handler {
@@ -731,7 +721,7 @@
     _handler = [handler copy];
     [callbackMemory setValue:_handler forKey:hash];
     
-    [self executeJS:[NSString stringWithFormat:@"Mapwize.api.getPlaceList({name:'%@', venueId:'%@'}, function(err, placeList){map.fire('apiResponse', {returnedType:'placeList', hash:'%@', response:placeList, error:err?err.status:''});});", placeListName, venueId, hash ]];
+    [self executeJS:[NSString stringWithFormat:@"Mapwize.Api.getPlaceList({name:'%@', venueId:'%@'}, function(err, placeList){map.fire('apiResponse', {returnedType:'placeList', hash:'%@', response:placeList, error:err?err.status:''});});", placeListName, venueId, hash ]];
 }
 
 - (void) getPlaceListWithAlias: (NSString*) placeListAlias inVenue:(NSString*) venueId completionHandler:(void(^)(MWZPlaceList*, NSError*)) handler {
@@ -740,7 +730,7 @@
     _handler = [handler copy];
     [callbackMemory setValue:_handler forKey:hash];
     
-    [self executeJS:[NSString stringWithFormat:@"Mapwize.api.getPlaceList({alias:'%@', venueId:'%@'}, function(err, placeList){map.fire('apiResponse', {returnedType:'placeList', hash:'%@', response:placeList, error:err?err.status:''});});", placeListAlias, venueId, hash ]];
+    [self executeJS:[NSString stringWithFormat:@"Mapwize.Api.getPlaceList({alias:'%@', venueId:'%@'}, function(err, placeList){map.fire('apiResponse', {returnedType:'placeList', hash:'%@', response:placeList, error:err?err.status:''});});", placeListAlias, venueId, hash ]];
 }
 
 - (void) getPlaceListsForVenue:(NSString *)venueId completionHandler:(void (^)(NSArray *, NSError*))handler {
@@ -749,7 +739,7 @@
     _handler = [handler copy];
     [callbackMemory setValue:_handler forKey:hash];
     
-    [self executeJS:[NSString stringWithFormat:@"Mapwize.api.getPlaceLists('%@', function(err, placeLists){map.fire('apiResponse', {returnedType:'placeLists', hash:'%@', response:placeLists, error:err?err.status:''});});", venueId, hash ]];
+    [self executeJS:[NSString stringWithFormat:@"Mapwize.Api.getPlaceLists('%@', function(err, placeLists){map.fire('apiResponse', {returnedType:'placeLists', hash:'%@', response:placeLists, error:err?err.status:''});});", venueId, hash ]];
 }
 
 - (void) getPlacesWithPlaceListId:(NSString *)placeListId completionHandler:(void (^)(NSArray*, NSError*))handler {
@@ -758,28 +748,45 @@
     _handler = [handler copy];
     [callbackMemory setValue:_handler forKey:hash];
     
-    [self executeJS:[NSString stringWithFormat:@"Mapwize.api.getPlaces({placeListId:'%@'}, function(err, places){map.fire('apiResponse', {returnedType:'places', hash:'%@', response:places, error:err?err.status:''});});", placeListId, hash ]];
+    [self executeJS:[NSString stringWithFormat:@"Mapwize.Api.getPlaces({placeListId:'%@'}, function(err, places){map.fire('apiResponse', {returnedType:'places', hash:'%@', response:places, error:err?err.status:''});});", placeListId, hash ]];
 }
 
 - (void) refresh {
     [self executeJS:[NSString stringWithFormat:@"map.refresh()"]];
 }
 
+- (void) startLocationWithBeacons:(BOOL) useBeacons {
+    [self unlockUserPosition];
+    [_locationManager startUpdatingLocation];
+    [_locationManager startUpdatingHeading];
+    _options.beaconsEnabled = useBeacons;
+    if (_options.beaconsEnabled) {
+        [self updateMonitoring:monitoredUuids];
+    }
+}
+
+- (void) stopLocation {
+    [_locationManager stopUpdatingLocation];
+    [_locationManager stopUpdatingHeading];
+    [self setUserHeading:nil];
+    [self removeUserPosition];
+    _options.beaconsEnabled = false;
+    [self updateMonitoring:@[]];
+}
 
 /* location manager methods */
 - (void) updateMonitoring: (NSArray*) uuids {
-    
     NSArray* rangedRegion = [[_locationManager rangedRegions] copy];
     for (CLBeaconRegion* region in rangedRegion) {
         [_locationManager stopRangingBeaconsInRegion:region];
     }
-    
-    for (NSString* uuidString in uuids) {
-        NSUUID *uuid = [[NSUUID alloc] initWithUUIDString:uuidString];
-        CLBeaconRegion* region = [[CLBeaconRegion alloc] initWithProximityUUID:uuid identifier:uuidString];
-        [_locationManager startRangingBeaconsInRegion:region];
+    if (_options.beaconsEnabled) {
+        for (NSString* uuidString in uuids) {
+            NSUUID *uuid = [[NSUUID alloc] initWithUUIDString:uuidString];
+            CLBeaconRegion* region = [[CLBeaconRegion alloc] initWithProximityUUID:uuid identifier:uuidString];
+            [_locationManager startRangingBeaconsInRegion:region];
+        }
     }
-    
 }
 
 #pragma mark Location Manager Delegate
@@ -789,13 +796,17 @@
     [self newUserPositionMeasurement:m];
 }
 
-- (void)locationManager:(CLLocationManager *)manager
-       didFailWithError:(NSError *)error {
-    NSLog(@"%@", error);
+- (void)locationManager:(CLLocationManager *)manager didFailWithError:(NSError *)error {
+    
+}
+
+- (void)locationManager:(CLLocationManager *)manager didUpdateHeading:(CLHeading *)newHeading {
+    [self setUserHeading:[NSNumber numberWithFloat:newHeading.trueHeading]];
 }
 
 - (void)locationManager:(CLLocationManager *)manager didRangeBeacons:(NSArray<CLBeacon *> *)beacons inRegion:(CLBeaconRegion *)region {
     NSMutableArray* beaconsToJS = [[NSMutableArray alloc] init];
+    
     for (CLBeacon* beacon in beacons) {
         NSMutableDictionary* beaconDic = [[NSMutableDictionary alloc] init];
         [beaconDic setObject:[beacon.proximityUUID UUIDString] forKey:@"uuid"];
@@ -807,7 +818,7 @@
     }
     NSData *beaconsJSON = [NSJSONSerialization dataWithJSONObject:beaconsToJS options:(NSJSONWritingOptions) 0 error:nil];
     NSString* beaconsString = [[NSString alloc] initWithData:beaconsJSON encoding:NSUTF8StringEncoding];
-    [self executeJS:[NSString stringWithFormat:@"map.addRangedIBeacons(%@)", beaconsString]];
+    [self executeJS:[NSString stringWithFormat:@"Mapwize.Location.addRangedIBeacons('%@',%@);", region.identifier, beaconsString]];
 }
 
 @end